refactor: cleaned up files and setup structure so theres a components folder. Also, fixed warnings from CoursePage.js
